+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 ## HEAD (Unreleased)
 _(none)_
 
+- Fix an issue in the `ignoreChanges` diffing logic which prevented deep matching properties
+  which used a period in their path keys.
+
 ## 2.16.2 (2020-12-23)
 
 - Fix a bug in the core engine that could cause previews to fail if a resource with changes had